[FEAT] : 예약 취소 API 개발 / 마이페이지 예약 조회 API 개발  (#85)
* [FEAT] : 예약 취소 DTO 개발

* [FEAT] : 예약 취소 DTO 개발

* [FEAT] : 예약 취소 성공,실패 상태 추가

* [FIX] : 예약 상태 이름 변경

* [FEAT] : Booking 엔티티에 cancel 비지니스 메서드 추가

* [FEAT] : 예약 취소 API 구현

* [FEAT] : Booking 엔티티 취소 이유 필드와 취소 메서드 구현

* [FEAT] : BookingConverter 구현

* [FEAT] : 예약생성 응답 DTO에 paymentID, orderID 추가

* [CHORE] : 결제 요청 API 스웨거 설명 수정

* [REFACTOR] : 결제 승인 API에서 예약 상태도 변경하도록 코드 구조 변경

* [CHORE] : 결제 완료 처리 api 스웨거 설정 변경

* [REFACTOR] : 예약 생성 api에서 내부적으로 결제 요청 로직까지 진행하도록 코드 변경

* [FEAT] : 예약 엔티티에서 결제 완료된 결제키를 찾는 편의 메서드 개발

* [FEAT] : 예약 취소시 환불까지 연동되는 비즈니스로직 개발

* [FEAT] : 예약 조회 응답 DTO 개발

* [FEAT] : 예약 조회 Repository 개발

* [FEAT] : 예약 조회 Service 개발

* [FEAT] : 예약 조회 Controller 개발

* [CHORE] : application-local.yml 파일 설정 추가

* [FEAT] : BookingMenu 엔티티 개발

* [FEAT] : 예약 생성 요청 DTO 변경

* [FEAT] : 예약 생성 서비스 로직 변경(메뉴 고려)

* [FIX] : 예약금, 결제금액 타입 Integer->Decimal로 변경

* [FIX] : 예약 조회 기본 페이지 1로 변경

* [FIX] : 불필요한 예약 완료 처리 api 삭제

* [REFACTOR] : 존재하지 않는 테이블을 포함해 에약을 생성하는 예외 처리

* [CHORE] : PAYMENT 컨트롤러 SWAGGER 설정 복원

* [FEAT] : 예약 취소 로직 예외 처리 추가

* [FIX] : 결제 금액 타입 Integer -> BigDecimal로 변경

+++ src/main/java/com/eatsfine/eatsfine/domain/payment/dto/response/PaymentResponseDTO.java
@@ -1,6 +1,7 @@
 package com.eatsfine.eatsfine.domain.payment.dto.response;
 
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -10,7 +11,7 @@ public class PaymentResponseDTO {
             Long paymentId,
             Long bookingId,
             String orderId,
-            Integer amount,
+            BigDecimal amount,
             LocalDateTime requestedAt) {
     }
 
@@ -26,7 +27,7 @@ public class PaymentResponseDTO {
             Long paymentId,
             Long bookingId,
             String storeName,
-            Integer amount,
+            BigDecimal amount,
             String paymentType,
             String paymentMethod,
             String paymentProvider,
@@ -51,7 +52,7 @@ public class PaymentResponseDTO {
             String storeName,
             String paymentMethod,
             String paymentProvider,
-            Integer amount,
+            BigDecimal amount,
             String paymentType,
             String status,
             LocalDateTime requestedAt,
@@ -65,7 +66,7 @@ public class PaymentResponseDTO {
             String status,
             LocalDateTime approvedAt,
             String orderId,
-            Integer amount,
+            BigDecimal amount,
             String paymentMethod,
             String paymentProvider,
             String receiptUrl) {
